feat(export): journal/publication sizing — physical mm-width + DPI
Quick-Wins #3. Adds an opt-in physical sizing path to /export: when width_mm
AND dpi are both set, the output is sized to a physical width (target px =
width_mm/25.4 * dpi, a float upscale factor via PIL resize) and the dpi is
embedded in the PNG/JPEG/PDF, so figures import at the intended size in
Illustrator/Word. The integer scale path is unchanged and byte-identical
(eff_scale = float(scale); every overlay size is round(int*float(int)) = the
old int product). tiff16 stays integer-scale (quantitative data export).

ExportDialog gains a Scale/Physical toggle with Nature/Science/ACS column-
width presets and a 150/300/600 dpi picker; the overlay helpers in calc/
export.py now accept a float scale so scale bars/labels stay proportional at
any factor. 4 new pytest + 3 new vitest tests; 50 pytest + 260 vitest green.

+++ src/fermiviewer/calc/export.py
@@ -163,7 +163,7 @@ def _box_outline(
     kind: str,
     ipts: list[tuple[float, float]],
     opts: tuple[tuple[float, float], ...],
-    scale: int,
+    scale: float,
 ) -> list[Annotation]:
     """Box profiles (width set) bake the averaging-box outline around
     the dashed centerline — mirrors MeasureOverlay. Empty list for
@@ -188,7 +188,7 @@ def measure_annotations(
     img_w: int,
     pixel_size: float | None,
     pixel_unit: str,
-    scale: int,
+    scale: float,
     raster: np.ndarray | None = None,
     tilt_angle_deg: float = 0.0,
     tilt_axis: str = "Y",
@@ -321,7 +321,7 @@ def scale_bar_geometry(
     out_h: int,
     pixel_size: float,
     pixel_unit: str,
-    scale: int,
+    scale: float,
     *,
     norm_x: float | None = None,
     norm_y: float | None = None,
